refactor: move core/shared/lib into core/lib/acl and core/lib/log
