+++ core/lib/log/aggregate.test.ts
