+++ core/lib/log/aggregate.ts
