+++ core/lib/log/parse-identifier.ts
